Moved SecretHelper into Specs project as this is the only place it is used and don't want to expose injection of IConfiguration outside spec scenarios.
